# Add fast path for Parquet reading with predicate pushdown via AST filters