Closes #6004 

Benchmarks: TL;DR is a slight possibly negligible improvement (no regressions).
With this PR
```
------------------------------------------------------------------------------------------------ benchmark: 5 tests -----------------------------------------------------------------------------------------------
Name (time in ms)                                   Min                 Max                Mean            StdDev              Median               IQR            Outliers       OPS            Rounds  Iterations
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
bench_read_parquet_with_filters[10000]           6.3109 (1.0)        7.0931 (1.0)        6.4235 (1.0)      0.2258 (1.55)       6.3470 (1.0)      0.0270 (1.0)           1;2  155.6781 (1.0)          11           1
bench_read_parquet_with_filters[100000]          8.1858 (1.30)       8.7125 (1.23)       8.3167 (1.29)     0.1459 (1.0)        8.2445 (1.30)     0.1372 (5.08)        20;14  120.2394 (0.77)        103           1
bench_read_parquet_with_filters[1000000]        11.6107 (1.84)      13.0220 (1.84)      11.9153 (1.85)     0.2429 (1.67)      11.8608 (1.87)     0.2542 (9.41)         16;4   83.9255 (0.54)         77           1
bench_read_parquet_with_filters[10000000]       22.5830 (3.58)      25.9029 (3.65)      23.2476 (3.62)     0.7875 (5.40)      23.0457 (3.63)     0.4460 (16.51)         1;1   43.0152 (0.28)         15           1
bench_read_parquet_with_filters[100000000]     111.8225 (17.72)    116.4201 (16.41)    112.9419 (17.58)    1.4617 (10.02)    112.6221 (17.74)    0.7594 (28.11)         1;1    8.8541 (0.06)          8           1
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

Before this PR:
```
------------------------------------------------------------------------------------------------ benchmark: 5 tests -----------------------------------------------------------------------------------------------
Name (time in ms)                                   Min                 Max                Mean            StdDev              Median               IQR            Outliers       OPS            Rounds  Iterations
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
bench_read_parquet_with_filters[10000]           6.3168 (1.0)        7.1415 (1.0)        6.4278 (1.0)      0.2382 (1.74)       6.3538 (1.0)      0.0382 (1.0)           1;1  155.5753 (1.0)          11           1
bench_read_parquet_with_filters[100000]          8.1519 (1.29)       8.6667 (1.21)       8.2777 (1.29)     0.1367 (1.0)        8.2275 (1.29)     0.0701 (1.84)        14;14  120.8062 (0.78)        100           1
bench_read_parquet_with_filters[1000000]        11.4127 (1.81)      12.5740 (1.76)      11.5801 (1.80)     0.2028 (1.48)      11.5040 (1.81)     0.0871 (2.28)         8;11   86.3551 (0.56)         80           1
bench_read_parquet_with_filters[10000000]       22.7609 (3.60)      30.2208 (4.23)      23.6469 (3.68)     1.8468 (13.51)     23.1346 (3.64)     0.4219 (11.05)         1;2   42.2888 (0.27)         15           1
bench_read_parquet_with_filters[100000000]     114.8168 (18.18)    119.4892 (16.73)    117.0239 (18.21)    1.6934 (12.39)    116.5923 (18.35)    2.7417 (71.84)         4;0    8.5453 (0.05)          8           1
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

At `n=500_000_000` rows:

Before this PR:
```
------------------------------------------------------------ benchmark: 1 tests -----------------------------------------------------------
Name (time in ms)                                   Min       Max      Mean  StdDev    Median     IQR  Outliers     OPS  Rounds  Iterations
-------------------------------------------------------------------------------------------------------------------------------------------
bench_read_parquet_with_filters[500000000]     506.9252  515.4757  510.3525  3.3557  510.3305  4.6582       2;0  1.9594       5           1
-------------------------------------------------------------------------------------------------------------------------------------------
```


With this PR:
```
------------------------------------------------------------ benchmark: 1 tests -----------------------------------------------------------
Name (time in ms)                                   Min       Max      Mean  StdDev    Median     IQR  Outliers     OPS  Rounds  Iterations
-------------------------------------------------------------------------------------------------------------------------------------------
bench_read_parquet_with_filters[500000000]     502.8996  512.1174  508.1085  3.3424  508.1873  3.2819       2;0  1.9681       5           1
-------------------------------------------------------------------------------------------------------------------------------------------
```

## Description
<!-- Provide a standalone description of changes in this PR. -->
<!-- Reference any issues closed by this PR with "closes #1234". -->
<!-- Note: The pull request title will be included in the CHANGELOG. -->

## Checklist
- [ ] I am familiar with the [Contributing Guidelines](https://github.com/rapidsai/cudf/blob/HEAD/CONTRIBUTING.md).
- [ ] New or existing tests cover these changes.
- [ ] The documentation is up to date with these changes.